Tighten up scalar/vector value semantics.  Missing allele indices are now None.
Scalar types with vector values are now returned as vectors.  The
alternative is to make them inaccessible or to silently drop values.  It
_should_ be an error to write such values to file, but that is up to htslib
to handle.

Alleles and allele indices are vector types, so much always be returned as
tuples.  A few places returned None when genotypes were trivially unset,
rather than ().