Multiple fixes related to VariantRecordInfo and handling of INFO/END (#461)
A pysam user has privately reported a bug with the VariantRecord class.  The 
situation is somewhat complicated, but the root cause is that INFO/END is
able to get out of sync with the VariantRecord stop coordinate.  After
trying several simpler workarounds, it seems that the best solution is to
not expose INFO/END via VariantRecord.info.  This is because it is an
artifact of the VCF text format that isn't used in BCF or the htslib
in-memory representation.  Instead, all access is via VariantRecord.stop and
VariantRecord.rlen, while INFO/END is kept in sync behind the scenes.

In fixing researching and this bug, I've also found and fixed several other
bugs in VariantRecordInfo:
  * scalar INFO entries could not be set again after being deleted
  * multiple cases where access to values was attempted before ensuring
    that lazy parsing had occurred
  * __bool__ and __len__ counted deleted entries
  * __delitem__ now allows unset flags to be deleted without raising a KeyError