Refactor region_parsing and related code to HTSFile base class.
Doing so exposed a bifurcation in terminology.  AlignmentFile uses
`reference` and `end`, where as VariantFile is careful to always use
`contig` and `stop`.  Although there is more code in the wild using
AlignmentFile, I *strongly* prefer the latter nomenclature and updated all
of the code to use that.  However, extra keyword arguments were added to
enable backward compatible use of `reference` and `end`.  I understand this
change may be a bit controversial and I am happy to discuss before merging
this PR.