Fix bug in get_aligned_pairs
I  noticed that in pysam 0.15.1, there is a bug with the get_aligned_pairs function when matches_only=True and with_seq=True.  Basically, the read position gets advanced, but the reference position does not.  This fixes that bug, and I believe it won't involve changing any other portions of the code.