Update api.rst
When the window is a little bit larger, the original demo may raise error `exit code 139 (interrupted by signal 11: SIGSEGV)`. One assumed reason for this is the garbage collector in python deletes bamfile.pileup (...) while the loop is running and the code goes to empty memory. And to address this problem, the updated demo should be used.
Ref: https://sites.google.com/a/epigenetic.ru/lab-members/kb/it/pysam