Fix Travis build failures due to libdeflate; re-enable Travis lzma
In Travis builds, with the mixture of left-over conda CC/CFLAGS
settings now being used during configure, an HTSlib configure bug
was leading to build failures due to missing <libdeflate.h>.
This backports the HTSlib configure fix (cf samtools/htslib#903).

Re-enable lzma in Travis builds as the previous commit resolves
the macOS problem that led to it being disabled.

Make the start of the via-setup.py build easier to find in the logs,
and exit immediately if it fails.