Fix a bunch of documentation formatting errors
To fix a lot of these, I converted from what Napoleon calls "NumPy-style" docstrings to what it calls "Google-style" docstrings. I did this because I'm not actually sure how to get some of this stuff to render correctly in the NumPy style, and indeed think that in many cases it's impossible without making entire field descriptions be a single line. (I've opened https://github.com/sphinx-contrib/napoleon/issues/12 to point out one example where it seems to be impossible to write a NumPy-style docstring that Napoleon won't garble.)

There are some incidental stylistic changes that happened as side effects of these fixes (e.g. stuff changing between monospace and bold, or the type "bool" in `Optional[bool]` ceasing to be a link back to the Python standard library docs). That's not what I wanted, but none of the new versions are outright broken so I think it's fine.

This commit also includes some purging of trailing whitespace that my editor did when I saved.