Fix silent truncation of FASTQ with bad q strings
Prior to this commit, `pysam` silently truncates
FASTQ files that have bad quality strings. For instance,
consider the following code which attempts to read a FASTQ
file where the quality string is invalid because it differs
in length from the sequence:

    import tempfile
    import pysam

    with tempfile.NamedTemporaryFile('wt') as f:
        f.write('>name\nACT\n+\n????')
        f.flush()
        nreads = len([r for r in pysam.FastxFile(f.name)])
        print(nreads)

Prior to this commit, the above code would return 0 even
though there is in fact 1 invalid entry in the FASTQ.
The reason is that in the code, `kseq_read` returns -1
for end of file but -2 for truncated quality string, but
the code for `FastxFile` assumes end of file for any return
< 0.

With this commit, an error is now raised for invalid FASTQ
files rather than silently truncating the file.