Fix a broken import
Without this change, installing from the repo (with `pip3 install .` or `python3 setup.py install`) produces a broken installation that fails on this import when I try to do `import pysam`.

(I'm confused about why the version I get by running `pip3 install pysam` doesn't have the same issue. It seems to already have this change. Is there something in the build pipeline that rewrites this import, that's missing when I install locally? Or, @AndreasHeger, did you make this change locally before pushing to PyPI and not commit it? Sorry if this is a dumb question; I'm inexpert at how setuptools works.