Allow VariantFile.fetch() on files with no records (issue #904)
When there are no records, .tbi indexes contain no reference sequences,
which BaseIndex.__bool__() over-cleverly reports as False. In fetch(),
what we're really checking for is that self.index is an Index object,
not that it's an Index object that is True. (This is the only such
instance of `if [not] self.index`.)