Choose TBI/CSI in tabix_index() via both min_shift and csi
Only min_shift is passed to the HTSlib functions that actually make the
index. Thus if only csi is set by the caller, in all cases we need to set
min_shift > 0 to signal the request for CSI to the underlying function.
If min_shift is unset, set it to the default 14. Fixes #995.
Also update the documentation accordingly.

In tests, also check that the file written is in the expected TBI\1
or CSI\1 format. Previously test_vcf_with_csi_index() was incorrectly
writing TBI contents to a file named foo.csi.