Properly import libchtslib
This is only for type checking: Mypy (and PyCharm) complain that libchtslib
is undefined (within the definition of __all__).

This is actually not true because "from pysam.x import y", when used within
pysam/__init__.py, will actually add not only y, but also x into the list
of available symbols.