Add basic CI GitHub Actions workflow
Basic build (directly from the checked-out source code) and test on
a matrix of Python versions on Linux and macOS.

(Factoring "-latest" out of matrix.os shortens the annotated job names
so they aren't trimmed when displayed. If in future we need to pin to
e.g. macos-10.15 we will need to undo that and accept the trimming.)