Clarify get_forward_sequence/get_forward_qualities documentation
In the documentation for query_sequence and these two methods, spell out
that SEQ/QUAL are stored reverse complemented (compared to the original
unaligned read) for reads mapped to the reverse strand (per is_reverse).
Clarify that get_forward_sequence reverse complements the stored SEQ and
get_forward_qualities reverses the stored QUAL for these is_reverse reads
so as to return them in the original unaligned orientation.