Add decode_bytes() function and update optional argument handling
Add a new decode_bytes() function that decodes UTF-8 even in Python 2,
alongside the others in libcutils.pyx so that it too can use the new
error handler infrastructure. Use it in bcf_array_to_object() so that
these fields are decoded even in Python 2.

Redo all these functions' optional argument handling so that they
default to the *current* value of TEXT_ENCODING when omitted in a call.
Also add 'errors' optional arguments that default to ERROR_HANDLER.

Lift the registering of 'pysam.latin1replace' so that it always happens
even without calling set_encoding_error_handler(), ready for future
changes in which objects might have their own error handlers that they
pass to calls of these functions, instead of user code setting the global.