Add reimagined "import" subcommand to pysam/samtools.py
The older `samtools import` command, which was a simplified version of
view for format conversion, was removed in samtools-1.11. This samtools
version adds a new unrelated "import" command for importing FASTQ files,
hence this time from Python we call it "fqimport" (to avoid conflicting
with the reserved word) rather than the old "samimport".