Remove use_2to3 from setup.py metadata
As of v58.0.2 (released 6 September), setuptools immediately rejects
any use of use_2to3 -- especially setting it to True.

Fortunately we no longer need it anyway. The only change it currently
applies is wrapping `list()` around the following in pysam/samtools.py:

    for key, options in list(SAMTOOLS_DISPATCH.items()):

which is not desired/required anyway as SAMTOOLS_DISPATCH does not
change during the loop.