Refactor tabix_index() format detection
The is_gzip_file() function neglects to close fd, so leaks a file
descriptor. Rather than fix this directly, refactor the code so
that this check is done via the subsequent hts_get_format()-based
format checking, thus also reducing the number of times the file
is opened and closed. Fixes #1038.