Introduce TupleProxyIterator iterator object class
Zeroing self.index in __iter__() led to incorrect results when __iter__()
is subsequently called (on TupleProxy used as an iterator object) in code
such as

    for row in tabixfile.fetch(..., parser=pysam.asTuple()):
        contig, start, end, *others = row

Fix this by introducing a separate TupleProxyIterator class so that the
current field counter can be zeroed in TupleProxy.__iter__() but not
TupleProxyIterator.__iter__(). Fixes #1053.

Add test cases, including a new file (fivecolumns.bed) and Makefile rules
to compress and index it. Use a circumlocution via `exec()` to hide the
`*others` syntax error from Python 2.7 -- and look forward to the day
that we drop Python 2!