Update devtools/import.py: Avoid rewriting identifiers like "xyz_stdout"
Samtools contains various identifiers like "hstdout", which we were
harmlessly rewriting to "hsamtools_stdout". However the upcoming
samtools 1.14 contains a function, autoflush_if_stdout(), that is
declared in samtools.h, which is not rewritten to samtools.h.pysam.h.
Rewriting references to it in *.c.pysam.c will cause build failures,
so tighten up the regexps so that stdout/stderr is only rewritten
when it is the entire identifier.

Samtools 1.14's sam_utils.c contains its own samtools_stdout variable,
which collides with the generated one in samtools/samtools.pysam.h.
So unfortunately we must rewrite the sam_utils.c one to a different
name first.