Remove generated files in tests/tabix_data
Remove *.tbi files that can be regenerated with tabix and *.gz files
that can be regenerated with bgzip, and add them as targets to the
Makefile instead.

Leave example_0v* as is, as they intentionally test index files
generated by old versions of tabix. Leave some *.gz files as is,
as they are not reproduced exactly by current bgzip.

Leave example.gtf.gz.tbi as is, as we need to ensure it was generated
with HTSlib 1.13 or later (cf c8ab144d263d) and in particular CI hosts
may provide an earlier version.