Alter function names via #define rather than rewriting
Several non-static function names are used in both samtools and bcftools.
Both libcsamtools.so and libcbcftools.so are loaded simultaneously, leading
to collisions and wrong functions being called. #define these names so the
actual symbol names include distinct prefixes to avoid collisions.

Some of the symbols we will want to rename appear in headers as well as
C files, and the rewriting mechanism is currently applied only to C files.
Hence the current rewriting infrastructure would not suffice.

(Remove obsolete comment -- we do rename `main_reheader`.)