Don't check for symbol conflicts when HTSLIB_MODE=separate
In separate build mode, all the extension modules will contain most
of the htslib symbols. Hence this check always fails in this mode.

The check is more relevant for shared mode. Fortunately nobody seems
to have encountered this problem since the check was added in February,
and frankly separate mode should perhaps be removed.