Simplify `pysam.{sam,bcf}tools.index(...)` argument checking
As a special case and so that it can give an obvious clear exception,
_pysam_dispatch checks that the (first) input file given to `index`
exists (as a local file). Since PR #755, this contains a duplicate
simplified option parser so that arguments that are options are ignored.
This additional option parser has not been updated as options have been
added to the underlying `samtools/bcftools` index commands.

Simplify this to only explicitly list the underlying options that have
arguments themselves -- we don't need to track simple flags explicitly
to skip them. Fixes #1155.