Ignore CIGAR-less reads in AlignmentFile.find_introns
This function iterates over read.cigartuples, which failed when that
was None. It's safe to just skip such reads, as they have no N operations
so don't contribute to the count anyway. Fixes #1148.

Add a test case containing an unmapped read with no CIGAR string and adjust
test_total accordingly. This needs to be a placed unmapped read, as the
test cases fetch only reads with RNAME set, not truly unmapped reads.