Add sysconfig's CCSHARED to environmentally supplied CFLAGS
Users who set CFLAGS via the environment may neglect to add appropriate
PIC options, so we should add CCSHARED to the value provided, just as we
do when CFLAGS is provided via sysconfig. Fixes #1146.

(In future when the configure step is moved to prebuild_libchtslib,
it will be able to use build_config_dict and set_compiler_envvars can
be removed. For now, we need this hack.)