Re-add Python 3.11 to the CI matrix and suppress flaky test
Test on 3.11 since it has been released for a while. The macOS py-3.11
build fails due to a mysterious missing liblzma, so (for now) activate
only py-3.10 on the macOS "direct" job, not 3.11.

Disable testIteratorOutOfScope test on Python 3.11. This test exercises
invalid code and checks that it results in an exception rather than a
segfault. It seems that Python 3.11 has changed the way unreferenced
objects call Cython deallocators, or something similar, and the code
leads to a crash on 3.11. Work around #1151 by disabling the test
for the time being.