Fix macOS Python 3.11 build by removing brewed liblzma
Python 3.11 is now the first actions/setup-python Python built
to target universal2 (i.e., x86_64/arm64 multiarch) on macOS.

Brewed packages such as xz (which provides liblzma.dylib and liblzma.a)
are built for only a single native architecture. This led to build and
configure failures when the native-only liblzma.dylib was linked against
by a multiarch compilation.

Work around this by unlinking the xz package so that this brewed library
is not found by the linker, failing back to the system .tbd file instead.
(Ignore errors, in case there are scenarios in which it is not installed.)
If later build steps use brewed commands that need this library, we may
need to `brew link` it again later, but for now we get away without it.