Use HTS_IDX_SILENT_FAIL when loading index
Suppresses an HTSlib diagnostic when AlignmentFile._open() checks for
an index file even if the caller doesn't request it, in which case an
absent index file is not an error. Fixes #939.

Add HTS_IDX_SAVE_REMOTE, HTS_IDX_SILENT_FAIL, and the *_load3() functions
that use them to libchtslib.pxd.