Remove obsolete setup() `requires` entry (adapted from PR #1216)
Unlike install_requires and setup_requires, it seems that this no longer
really does anything. It adds a `Requires: cython (>=0.29.12)` line to
the resulting wheel's pysam-x.y.z.dist-info/METADATA file, but that's
a v1.1 item that was superseded in 2005. As pyproject.toml specifies
the requirement anyway, we can just delete this.

Now that we depend on settings in pyproject.toml, update the setuptools
requirement to v61.0.0, which is documented as the version that introduced
robust pyproject.toml support.

The distutils.errors fallback was already unnecessary with a requirement
for setuptools >= v59.0.0, so remove it.

Co-Authored-By: John Marshall <jmarshall@hey.com>