Limit current pysam to Cython 3.x avoiding eventual 4.0 (PR #1218)
If there is ever a Cython 4 release, the major version number bump will
be due to some incompatible change. While we fully plan to support such a
Cython version in the contemporaneous future pysam, today's releases may
not be compatible with it. Mark them as such -- to be revisited during
such a Cython's development cycle and pre-releases. Fixes #1217.