Update package description metadata
Wordsmith the long description, add CRAM to the list of formats,
move the URLs to project.urls so they appear in the PyPI sidebar,
and specify the content type (to avoid twine warnings).

Wordsmith the (short) description and ensure it is the same in setup()
and pyproject.toml. Make pyproject.toml's version entry dynamic instead
of duplicating the canonical value in pysam/version.py.