Clean libhts.a between wheels and simplify URL
Add before-build command to clean htslib/*.o and libhts.a before building
each wheel, as these are in the source directory and would otherwise be
reused (possibly for a different architecture!). Eventually we will
configure htslib out-of-tree within build_temp or perhaps build_platlib,
i.e., under build/, and will be able to remove this as build/ is known
to the Python infrastructure and automatically managed.

Simplify documentation URL, as we have switched our RTD default from
latest to stable.