Remove hts_probe_cc.sh and don't try to run it from Makefile
Pysam always runs configure, so we don't need or want this.
It was being run by `make distclean` on an otherwise clean directory
(and failing from an sdist, which didn't include hts_probe_cc.sh),
so use `-include` to avoid trying to rebuild it in this case.
Hat tip @nsoranzo.