Ensure only one `make -C pysam_data` etc can run at once
It's uncommon, but some people run the tests in parallel. This breaks
if multiple makes run in the same data directory at once, so we attempt
to prevent that (for most users, perhaps not on NFS...) via the usual
mkdir-based directory locking protocol. If it goes wrong, users can
clear the lock directories with the existing `make clean` target.