Fix wheels' bundled libcurl.so incompatibility with Debian [HACK]
Linux wheels are (currently) built on AlmaLinux, so the libcurl.so
bundled into the wheel follows Alma/Red Hat/Fedora conventions for the
location of its certificate bundle. This fails when the wheel is used on
a Debian/Ubuntu platform with a different convention for this location.
When not overridden by $CURL_CA_BUNDLE, work around this by specifying
the expected Debian bundle location if the Red Hat one isn't present.

Fixes #1257 via a hacky work around.

(Take care to preserve this patch when importing HTSlib in future.)