Don't modify VariantHeader.new_record()'s mutable arguments
Just process GT normally, as the underlying HTSlib bcf_update_format()
method itself handles GT specially to put it first. If it didn't, the
correct place to do this would be down in VariantRecordSample.update(),
as that is public too, perhaps with code here to handle GT carefully
when both kwargs and samples are present. Fixes #1308.

Verified that new_record()'s `info` argument is never modified. Added
a test case calling new_record() repeatedly with the same list[dict].