Update CIGAR_OPS and SAM_FLAGS type stub syntax
This avoids 'error: Detected enum "pysam.libcalignedsegment.CIGAR_OPS" in
a type stub with zero members. There is a chance this is due to a recent
change in the semantics of enum membership. If so, use `member = value`
to mark an enum member, instead of `member: type`'.

See https://typing.readthedocs.io/en/latest/spec/enums.html#defining-members