Move AlignedSegment cache fields to a new sub-object
Adding the query_qualities_str and query_alignment_qualities_str fields
changed the size of this class, which broke binary compatibility for
Cython projects using pysam. Restore this by rearranging the fields
so that sizeof(AlignedSegment) is restored to 72.

Add class size tests to prevent a repeat of this. We will in future
add other classes to these tests, which for now just verify the basic
AlignmentFile and VariantFile classes.