Deprecate future opaque structs and rename member structs/unions
We intend to make these structs opaque in future: their fields are
present in the C declarations for implementation reasons but should
not be used by user code, so there is no reason why the corresponding
Cython declarations should not be as incomplete structs.

Cython needs named types for some fields that can be anonymous structs
in C. Rename some of these pysam-local structs so that they follow
hts_opt_val_union's pattern and avoid polluting the namespace.