Update release workflows for Python 3.14, Intel macOS runner, etc
Add Python 3.14 to release workflows and update cibuildwheel to a release
that uses 3.14-final. Use the new macos-15-intel image instead of macos-13,
which is obsolete.

Reorganise the build matrix to separate the manylinux and musllinux jobs
so they run in parallel.

Move all wheel building to GitHub Actions: Cirrus CI's macos-sonoma-base's
default Python can no longer run current cibuildwheel (which requires 3.11),
so it is time to move the ARM wheels to be built on GitHub Actions, now that
that supports both Linux and macOS on ARM.