Add Python 3.14 to CI and test most Python versions on macOS
No longer need to be miserly with macOS actions runners, but continue to
suppress 3.8 to avoid a spurious error annotation. Update Python versions
used for sdist and conda CI tests, and use newish setup-miniconda facility
to suppress the defaults channel and avoid another annotation.

Build the release sdist on the latest Python release.